+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add Course filter ([#67](https://gitlab.com/Skn0tt/EntE/issues/67))
 - Add Report for whole grad year ([#232](https://gitlab.com/Skn0tt/EntE/issues/232))
 
 ## [v0.22.2] - 2019-04-30